+++ CHANGELOG.md
@@ -13,6 +13,7 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ### Changed
 
 ### Fixed
+- `git::Repo::oldest_common_descendant_on_current_branch`: Add missing intersection of descendants.
 
 
 ## 0.3.0